добавлено первое отношение м:м между жанрами и пользователями, даже по рест апи и снова bugfix

+++ src/main/java/com/pip_coursework/multipleKeys/usersGenresKey.java
@@ -2,6 +2,6 @@ package com.pip_coursework.multipleKeys;
 import java.io.Serializable;
 
 public class usersGenresKey implements Serializable {
-    private int userId;
-    private int genreId;
+    private long userId;
+    private long genreId;
 }
